# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Quin Travel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/tour/create": {
            "post": {
                "tags": [
                    "Tour"
                ],
                "summary": "Create a new tour",
                "description": "Returns new tour information",
                "operationId": "CreateTour",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "thumnail"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title of tour",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "thumnail": {
                                        "description": "Thumbnail of tour",
                                        "type": "file",
                                        "format": "file"
                                    },
                                    "images[]": {
                                        "description": "List images (max 4)",
                                        "type": "array",
                                        "items": {
                                            "type": "file",
                                            "format": "file"
                                        }
                                    },
                                    "type": {
                                        "description": "Type of tour",
                                        "default": "inside",
                                        "enum": [
                                            "inside",
                                            "outside"
                                        ]
                                    },
                                    "category": {
                                        "description": "Category of tour",
                                        "default": "tour",
                                        "enum": [
                                            "tour",
                                            "hotel"
                                        ]
                                    },
                                    "price": {
                                        "description": "Price of tour",
                                        "type": "integer",
                                        "default": 0
                                    },
                                    "country_id": {
                                        "description": "country of tour",
                                        "type": "integer",
                                        "default": 232
                                    },
                                    "price_child": {
                                        "description": "Price for children of tour",
                                        "type": "integer",
                                        "default": 0
                                    },
                                    "price_baby": {
                                        "description": "Price for baby of tour",
                                        "type": "integer",
                                        "default": 0
                                    },
                                    "percent_sale": {
                                        "description": "Percent sale of tour",
                                        "type": "integer",
                                        "default": 0
                                    },
                                    "additional_fee": {
                                        "description": "Additional_fee of tour",
                                        "type": "string",
                                        "default": "0"
                                    },
                                    "province_start_id": {
                                        "description": "Province start  of tour",
                                        "type": "integer"
                                    },
                                    "province_end_id": {
                                        "description": "Province end  of tour",
                                        "type": "integer"
                                    },
                                    "number_of_day": {
                                        "description": "Number day of  tour",
                                        "type": "integer",
                                        "default": 1
                                    },
                                    "tour_pakage": {
                                        "description": "Tour parkage",
                                        "default": "standard",
                                        "enum": [
                                            "luxury",
                                            "standard",
                                            "affordable",
                                            "saving"
                                        ]
                                    },
                                    "quantity": {
                                        "description": "Quantity of  tour",
                                        "type": "integer",
                                        "default": 1
                                    },
                                    "date_start": {
                                        "description": "Number day of  tour",
                                        "type": "string",
                                        "format": "date-time",
                                        "default": "2024-06-29 16:19:10"
                                    },
                                    "transportation": {
                                        "description": "Type transportation of  tour",
                                        "default": "airplane",
                                        "enum": [
                                            "airplane",
                                            "car"
                                        ]
                                    },
                                    "time_start": {
                                        "description": "Time start to leave  of  tour",
                                        "type": "string",
                                        "default": "08:30"
                                    },
                                    "date[]": {
                                        "description": "Process date of each day",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "title_process[]": {
                                        "description": "Process Title of each day",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    },
                                    "content[]": {
                                        "description": "Process content of each day",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "description": "JWT Token (Bearer <token>)"
            }
        }
    }
}